Renamed demo_dataset to load_demo
load_demo_dataset is a function, not a class. Renamed to represent an action
