+++ src/ml_tooling/data/__init__.py
@@ -1,6 +1,6 @@
 from .file import FileDataset
 from .sql import SQLDataset
 from .base_data import Dataset
-from .demo_dataset import load_demo_dataset
+from .load_demo import load_demo_dataset
 
 __all__ = ["FileDataset", "SQLDataset", "Dataset", "load_demo_dataset"]
